pyproject.toml
[project]
name = "final_price"
version = "0.1.0"
description = ""
authors = [
    { name = "Artsiom Zhvanko", email = "azhvanko@outlook.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (==0.118.0)",
    "gunicorn (==23.0.0)",
    "httptools (==0.6.4)",
    "itsdangerous (==2.2.0)",
    "orjson (==3.11.3)",
    "phonenumbers (==9.0.15)",
    "psycopg[c] (==3.2.10)",
    "pydantic-settings (==2.11.0)",
    "redis (==6.4.0)",
    "rq (==2.6.0)",
    "setproctitle (==1.3.7)",
    "sqladmin (==0.21.0)",
    "sqlalchemy (==2.0.43)",
    "structlog (==25.4.0)",
    "uvicorn-worker (==0.4.0)",
    "uvloop (==0.21.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
